Passing data spec.
